+++ packages/ui/src/Tooltip/index.tsx
@@ -8,7 +8,7 @@ import React, {
 import classNames from 'classnames';
 import TooltipContent from './components/TooltipContent';
 import TooltipToggler from './components/TooltipToggler';
-import {useClickOutside} from '@eruditorgroup/profi-toolkit';
+import {useClickOutside, useCombinedRef} from '@eruditorgroup/profi-toolkit';
 import type {ForwardRefExoticComponent, HTMLAttributes} from 'react';
 
 import styles from './Tooltip.module.scss';
@@ -56,9 +56,9 @@ const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
     ref,
   ) => {
     const [opened, setOpened] = useState(false);
-    const _ref = useRef<HTMLDivElement | null>(null);
+    const [combinedRef, setRef] = useCombinedRef<HTMLDivElement | null>(ref);
 
-    useClickOutside(_ref, () => {
+    useClickOutside(combinedRef, () => {
       if (trigger === 'click') {
         opened && setOpened(false);
       }
@@ -77,11 +77,7 @@ const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
             if (onMouseLeave) onMouseLeave(ev);
           }}
           {...props}
-          ref={(el) => {
-            _ref.current = el;
-            if (typeof ref === 'function') ref(el);
-            else if (ref) ref.current = el;
-          }}
+          ref={setRef}
         />
       </TooltipContext.Provider>
     );